Backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
azure-data-tables==12.4.4
openpyxl==3.1.3
gunicorn==22.0.0
